requirements.txt
protobuf
Flask
Werkzeug
dotenv
python-dotenv
langchain
google-generativeai
openai
pillow
pdf2image
